fix(administration): fix missing assignment of firstname, lastname and email (#521)
* fix(administration): cleanup transfer records fixing missing assignment of firstname, lastname and email #520
* adjust end2end test
