+++ tests/endtoend/NotificationInitScenario/ModifyCoreUserRoleScenario.cs
@@ -18,6 +18,7 @@
  ********************************************************************************/
 
 using FluentAssertions;
+using Org.Eclipse.TractusX.Portal.Backend.Administration.Service.Models;
 using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.DBAccess.Models;
 using Org.Eclipse.TractusX.Portal.Backend.PortalBackend.PortalEntities.Enums;
 using RestAssured.Response.Logging;
@@ -189,6 +190,14 @@ public class ModifyCoreUserRoleScenario : EndToEndTestBase
         if (companyUserDetails is null)
         {
             throw new Exception($"Could not get company user details from {endpoint} should not be null.");
+        }
+        if (companyUserDetails.FirstName is null)
+        {
+            throw new Exception($"Company user details from {endpoint} should not return firstName null");
+        }
+        if (companyUserDetails.LastName is null)
+        {
+            throw new Exception($"Company user details from {endpoint} should not return lastName null");
         }
         _username = companyUserDetails.FirstName + " " + companyUserDetails.LastName;
         return companyUserDetails.CompanyUserId.ToString();
